Update to stop services for Windows 10 only

+++ scripts/PrepareWindows.ps1
@@ -100,29 +100,31 @@ if ([bool](get-Command -Name 'servermanager.exe' -ErrorAction SilentlyContinue) 
     Get-ScheduledTask -TaskName 'ServerManager' | Disable-ScheduledTask | Out-Null
 }
 
-Write-Output "Disabling non-essential Windows services."
-'ajrouter', 'alg', 'appmgmt', 'appxsvc', 'bthavctpsvc', 'bits', 'btagservice', 'bthserv',
-    'bluetoothuserservice_*', 'peerdistsvc', 'captureservice_*', 'certpropsvc', 'nfsclnt',
-    'diagtrack', 'IpxlatCfgSvc', 'iphlpsvc', 'SharedAccess', 'irmon', 'vmicvss', 'vmictimesync',
-    'vmicrdv', 'vmicvmsession', 'vmicheartbeat', 'vmicshutdown', 'vmicguestinterface', 'vmickvpexchange',
-    'HvHost', 'lfsvc', 'MapsBroker', 'dmwappushsvc', 'PhoneSvc', 'SEMgrSvc', 'WpcMonSvc', 'CscService',
-    'NcdAutoSetup', 'Netlogon', 'NaturalAuthentication', 'SmsRouter', 'MSiSCSI', 'AppVClient', 'SNMPTRAP',
-    'SCPolicySvc', 'ScDeviceEnum', 'SCardSvr', 'SensorService', 'SensrSvc', 'SensorDataService', 'RetailDemo',
-    'icssvc', 'WMPNetworkSvc', 'wisvc', 'wcncsvc', 'FrameServer', 'WFDSConSvc', 'WebClient', 'TabletInputService',
-    'XboxNetApiSvc', 'XblGameSave', 'XblAuthManager', 'xbgm', 'WwanSvc', 'wuauserv' | ForEach-Object {
-    Stop-Service -Name $_ -PassThru -ErrorAction SilentlyContinue | Set-Service -StartupType Disabled -ErrorAction SilentlyContinue
-}
+if ($osData.Name -like "*Windows 10*") {
+    Write-Output "Disabling non-essential Windows services."
+    'ajrouter', 'alg', 'appmgmt', 'appxsvc', 'bthavctpsvc', 'bits', 'btagservice', 'bthserv',
+        'bluetoothuserservice_*', 'peerdistsvc', 'captureservice_*', 'certpropsvc', 'nfsclnt',
+        'diagtrack', 'IpxlatCfgSvc', 'iphlpsvc', 'SharedAccess', 'irmon', 'vmicvss', 'vmictimesync',
+        'vmicrdv', 'vmicvmsession', 'vmicheartbeat', 'vmicshutdown', 'vmicguestinterface', 'vmickvpexchange',
+        'HvHost', 'lfsvc', 'MapsBroker', 'dmwappushsvc', 'PhoneSvc', 'SEMgrSvc', 'WpcMonSvc', 'CscService',
+        'NcdAutoSetup', 'Netlogon', 'NaturalAuthentication', 'SmsRouter', 'MSiSCSI', 'AppVClient', 'SNMPTRAP',
+        'SCPolicySvc', 'ScDeviceEnum', 'SCardSvr', 'SensorService', 'SensrSvc', 'SensorDataService', 'RetailDemo',
+        'icssvc', 'WMPNetworkSvc', 'wisvc', 'wcncsvc', 'FrameServer', 'WFDSConSvc', 'WebClient', 'TabletInputService',
+        'XboxNetApiSvc', 'XblGameSave', 'XblAuthManager', 'xbgm', 'WwanSvc' | ForEach-Object {
+        Stop-Service -Name $_ -PassThru -ErrorAction SilentlyContinue | Set-Service -StartupType Disabled -ErrorAction SilentlyContinue
+    }
 
-Write-Output "Stopping OneDrive."
-Get-Process -Name OneDrive* | Stop-Process -ErrorAction SilentlyContinue 
+    Write-Output "Stopping OneDrive."
+    Get-Process -Name OneDrive* | Stop-Process -ErrorAction SilentlyContinue 
 
-if (Get-Command -Command 'Get-AppxPackage') {
-    '*3dbuilder*', '*alarms*', '*appconnector*', '*appinstaller*', '*communicationsapps*', '*calculator*', '*camera*',
-        '*feedback*', '*officehub*', '*getstarted*', '*skypeapp*', '*zunemusic*', '*zune*', '*maps*', '*messaging*',
-        '*solitaire*', '*wallet*', '*connectivitystore*', '*bingfinance*', '*bing*', '*zunevideo*', '*bingnews*', 
-        '*onenote*', '*oneconnect*', '*mspaint*', '*people*', '*commsphone*', '*windowsphone*', '*phone*', '*photos*',
-        '*bingsports*', '*sticky*', '*sway*', '*3d*', '*soundrecorder*', '*bingweather*', '*holographic*',
-        '*windowsstore*', '*xbox*' | ForEach-Object {
-        Get-AppxPackage -Name $_ -AllUsers -ErrorAction SilentlyContinue | Remove-AppxPackage -Package $_ -ErrorAction SilentlyContinue
+    if (Get-Command -Command 'Get-AppxPackage') {
+        '*3dbuilder*', '*alarms*', '*appconnector*', '*appinstaller*', '*communicationsapps*', '*calculator*', '*camera*',
+            '*feedback*', '*officehub*', '*getstarted*', '*skypeapp*', '*zunemusic*', '*zune*', '*maps*', '*messaging*',
+            '*solitaire*', '*wallet*', '*connectivitystore*', '*bingfinance*', '*bing*', '*zunevideo*', '*bingnews*', 
+            '*onenote*', '*oneconnect*', '*mspaint*', '*people*', '*commsphone*', '*windowsphone*', '*phone*', '*photos*',
+            '*bingsports*', '*sticky*', '*sway*', '*3d*', '*soundrecorder*', '*bingweather*', '*holographic*',
+            '*windowsstore*', '*xbox*' | ForEach-Object {
+            Get-AppxPackage -Name $_ -AllUsers -ErrorAction SilentlyContinue | Remove-AppxPackage -Package $_ -ErrorAction SilentlyContinue
+        }
     }
 }
